rename and move some utils
